Refactored API versioning examples into spring-boot-mvc-5, fixed imports and POM
