+++ spring-boot-modules/spring-boot-mvc-5/src/main/java/com/baeldung/versioning/ApiVersioningDemoApplication.java
@@ -1,4 +1,4 @@
-package com.baeldung.example.apiversioning;
+package com.baeldung.versioning;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
