+++ spring-boot-modules/spring-boot-mvc-5/src/main/java/com/baeldung/versioning/controller/UserParamController.java
@@ -1,6 +1,6 @@
 package com.baeldung.versioning.controller;
 
-import com.baeldung.example.apiversioning.model.UserV2;
+import com.baeldung.versioning.model.UserV2;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
