+++ spring-boot-modules/spring-boot-mvc-5/src/main/java/com/baeldung/versioning/controller/header/UserHeaderController.java
@@ -1,8 +1,8 @@
 // src/main/java/.../controller/header/UserHeaderController.java
 package com.baeldung.versioning.controller.header;
 
-import com.baeldung.example.apiversioning.model.UserV1;
-import com.baeldung.example.apiversioning.model.UserV2;
+import com.baeldung.versioning.model.UserV1;
+import com.baeldung.versioning.model.UserV2;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RestController;
