+++ spring-boot-modules/spring-boot-mvc-5/src/main/java/com/baeldung/versioning/controller/mime/UserMimeController.java
@@ -1,8 +1,8 @@
 // src/main/java/com/baeldung/example/apiversioning/controller/mime/UserMimeController.java
 package com.baeldung.versioning.controller.mime;
 
-import com.baeldung.example.apiversioning.model.UserV1;
-import com.baeldung.example.apiversioning.model.UserV2;
+import com.baeldung.versioning.model.UserV1;
+import com.baeldung.versioning.model.UserV2;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
